Sensores

+++ solution/Evolvers/Neural/NeuralLayer.cs
@@ -9,12 +9,12 @@ namespace Evolvers.Neural
     {
         public IEnumerable<Neuron> Neurons { get; private set; }
 
-        public NeuralLayer(float[][] weights)
+        public NeuralLayer(double[][] weights)
         {
             Neurons = Enumerable.Range(0, weights.Count()).Select(i => new Neuron(weights[i])).ToArray();
         }
 
-        internal IEnumerable<float> GetSignals(IEnumerable<float> inputs)
+        internal IEnumerable<double> GetSignals(IEnumerable<double> inputs)
         {
             return Neurons.Select(n => n.GetSignal(inputs)).ToArray();
         }
